+++ tests/test_00_setup_ui.py
@@ -1,5 +1,5 @@
 import pytest
-from sh import temboard, temboard_migratedb, ErrorReturnCode
+from sh import temboard, ErrorReturnCode
 
 
 def test_temboard_version():
@@ -19,11 +19,11 @@ def test_temboard_help():
 
 def test_migratedb_no_database():
     with pytest.raises(ErrorReturnCode):
-        temboard_migratedb('check')
+        temboard('migratedb', 'check')
 
 
 def test_auto_configure(ui_auto_configure):
-    temboard_migratedb('check')
+    temboard('migratedb', 'check')
 
 
 def test_login_logout(browser, ui, ui_url):